chatbot-portfolio/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
faiss-cpu
sentence-transformers
